Tests/LibWeb/Text/expected/FileAPI/Blob-partitioning.txt

TypeError: Failed to obtain a Blob object from 'blob:' URL
